Added JUnit tests
